Moved common workflow execution controller actions into a concern, updated listing to only allow the correct user to delete/cancel, updated tests
